replaced volume binding with mount binding
The difference between them: https://docs.docker.com/storage/bind-mounts/#differences-between--v-and---mount-behavior

Because the -v and --volume flags have been a part of Docker for a long time, their behavior cannot be changed. This means that *there is one behavior that is different between -v and --mount*.

If you use -v or --volume to bind-mount a file or directory that does not yet exist on the Docker host, -v creates the endpoint for you. *It is always created as a directory*.

If you use --mount to bind-mount a file or directory that does not yet exist on the Docker host, Docker does not automatically create it for you, but generates an error.

This should save our SNO on Windows and MacOS from losing data during bugs in docker itself, when it doesn't map the data folder from the host to the container and just creates the data mountpoint without an error.