Add newline at EOF for Markdown files (#514)
Closes #513
POSIX requires text files to end in a line break. This is sometimes
assumed by some tools.

There was a mix of compliant and non-compliant markdown files,
suggesting there is no tool dependency on non-compliant files.

Useful scripts, credit to [this stackoverflow
answer](https://stackoverflow.com/a/77183929)
```
for f in $(
    grep --binary-files without-match --dereference-recursive --files-with-matches '' .
) ; do
    printf '%s: ' "$f"
    if test "$(tail --bytes 1 "$f")" ; then
        printf "does not end with line break\n"
    else
        printf "ends with line break\n"
    fi
done | grep 'does not end with line break'
```

```
for f in $( cat list.txt ) ; do
    printf '%s: ' "$f"
    echo "" >> "$f"
done | grep 'does not end with line break'
```

Co-authored-by: Ivan Fraixedes <ivan@fraixed.es>